requirements.txt
streamlit
openai
chromadb
pymupdf
python-docx
tavily-python
numpy
requests